requirements.txt
Flask==3.1.0
python-socketio==5.13.0